♻️ refactor: create reusable styled component for Form and Field
