+++ src/commons/components/FormWrapper.jsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const FormWrapper = styled.div({
+  display: 'flex',
+  flexDirection: 'column',
+  padding: '.5rem',
+});
+
+export default FormWrapper;
