+++ src/commons/components/index.js
@@ -1,0 +1,2 @@
+export { default as Button } from './Button';
+export { default as FormWrapper } from './FormWrapper';
